Remove unnecessary null pointer tests to avoid a false Coverity FORWARD_NULL defect.
In the for loop, at most one buffer in the |buffers| array
is of type SECBUFFER_DATA and at most one is of type SECBUFFER_EXTRA.
The null pointer test for decrypted_ptr_ is not necessary because
when we encounter the only buffer of type SECBUFFER_DATA, decrypted_ptr_
must be NULL (the initial value).  I added a DCHECK to assert that.
Similarly the null pointer test for received_ptr_ is not necessary.

R=mattm
BUG=22286
CID=5331
TEST=one
Review URL: http://codereview.chromium.org/371014

git-svn-id: http://src.chromium.org/svn/trunk/src@31499 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8852d6bc229045226562315f1597f3cdf7d4e5e9