Disabled extensions should be uninstallable.
The code to display the dialog looks up an
Extension by id but doesn't return disabled ones.

So I added that option to the function. 

I also added a check so that we don't show the
Options button enabled when the extension is
disabled, since you need the extension to be
enabled for the Options page to work.

BUG=26910
TEST=Disable an extension that has an Option page,
make sure Options is not enabled after disabling,
then uninstall it. It should uninstall.

Review URL: http://codereview.chromium.org/385014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@31612 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3f7d0a086d67dd00dac4936f47c14e689f4b05e0