Add a RWH_TabSwitchPaintDuration histogram.
The RWH_TabSwitchPaintDuration histogram calculates
the amount of time it took after a particular view was
selected for it to be fully painted.

I modified the tab switching test to use the new tab
switch paint duration histogram in place of the whiteout
duration histogram.  As a measure, whiteout duration
would not carry a value if we kept the backing store
around for a tab, and we would end up with a result of
zero instead of a useful result.  Local results with the
new measure are consistent with the number of operations
we perform on the tabs.

The tab switching test now fails if it cannot locate
the correct histogram.  In that case, it will output
the contents of the log file for remote debugging.

BUG=4104
TEST=count of tab switch paint duration histogram
results in tab switching test equals number of tabs
used in test

First committed as: http://src.chromium.org/viewvc/chrome?view=rev&revision=31632
Reverted in: 31645

Review URL: http://codereview.chromium.org/378038

git-svn-id: http://src.chromium.org/svn/trunk/src@31679 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 24a6b266d716505f6d6c0602124e21b6530f8e66