GTK: Use themed background images in toolbar buttons.
When a chrome theme asks to use a background image on a toolbar button,
use it. Oddly enough, all of the graphics drawing was written but the
plumbing of actually setting the background wasn't...

TEST=Good Smile, Chuck Anderson, etc. should have button backgrounds which
are very different from the toolbar background in the GTK port.
BUG=NONE

Review URL: http://codereview.chromium.org/384057

git-svn-id: http://src.chromium.org/svn/trunk/src@31728 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3a0c3a19e95a78e5eacf36e51a33eb7b55066e3f