linux: work around "flash does not register any clicks" bug
I had a flash of inspiration!  Setting GDK_NATIVE_WINDOWS in the
environment doesn't work because we don't propagate the setting down
into the plugin subprocess.  I confirmed that setting it manually
fixes it.  That is, on this computer I previously had the no-clicks
bug and with this change the problem is solved.

Note: there are multiple confusing levels to this bug; if
nspluginwrapper gets involved I still encounter the problem.
But this definitely fixes the problem for me.

BUG=20654

Review URL: http://codereview.chromium.org/384059

git-svn-id: http://src.chromium.org/svn/trunk/src@31734 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 38553160b3d58c9ed99c866cdac67e80dd290cdb