[Mac] Ignore incorrect focus notifications in Omnibox.
After the user runs a Print or Page Layout panel,
autocomplete_edit_view_mac.mm starts seeing
-controlTextDidBeginEditing: notifications when the field does not
have a field editor.  The code thoroughly depends on having a field
editor when it has focus, so this breaks things badly.
-controlTextDidBeginEditing: appears to be sent multiple times when in
this state, once inappropriately, and later when everything is set up
correctly (when the user edits), so this simply ignores the
inappropriate notification.

In the case that is being ignored, AutocompleteEditViewMac will fire
DCHECKs all over the place, so it's somewhat unlikely that this case
is happening in correct operation.

BUG=19116, 17803
TEST=Login to gmail, select a message, j, k, and other keys should work (and not send you to the omnibox).  Bring up the Print panel and cancel.  Bring up a new tab and close it.  gmail's accelerator keys should still work right (not send focus to omnibox).

Review URL: http://codereview.chromium.org/391001

git-svn-id: http://src.chromium.org/svn/trunk/src@31825 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8e11400b4b0a830f785c346c202c35900a431904