Always call NPP_SetWindow, even for 0x0 plugins, on the Mac
QuickTime can be embedded for audio, with a 0x0 plugin, but it still expects to get information about the window via SetWindow. Windows in fact made this same fix a while ago, for the same reason, but in the Windowed version of this method.

BUG=25883
TEST=Open a site like the one in the bug with a 0x0 QuickTime plugin; the plugin should not crash.

Review URL: http://codereview.chromium.org/389026

git-svn-id: http://src.chromium.org/svn/trunk/src@31845 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 41f1f582ae6d89796e6501bdbd1071a514c07d20