Add stub functions for chrome.* APIs in content scripts.
Most of the extensions API is not supported in content scripts for security 
purposes. Instead the typical pattern is to use messaging between content 
scripts and a background page which executes them on behalf of the script.
This is not immediately obvious to many extension developers, so this
change adds an error message to help them better understand what's going on.


BUG=26128
TEST=Try writing a content script that uses some of the extensions API stuff 
in chrome.tabs, chrome.windows, etc. You should see errors in the console 
telling you that it isn't supported in content scripts.


Review URL: http://codereview.chromium.org/366024

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@31958 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 98843c9af83df2d75a12382d03b44c4a1ae107ee