amit, please review everything, jam please review the changes to the tab_contents and the renderer_host sources.
Remove the AutomationProfileImpl class which wraps the Chrome profile for an external tab
container, which hosts ChromeFrame. This object was used to carry a custom URL request
context which was used to intercept HTTP requests and cookie requests issued by external
tabs.

However as the life time of the automation profile class depended on the lifetime of the external
tab container object this caused a number of crashes in objects which held on to the automation profile
pointer retrieved from the associated tab contents. This does not happen in a regualar Chrome browser 
instance as the profile is deleted at the very end.

We can associate the automation URL request context with the underlying tab_contents which would
eventually percolate down to the resource message filter. Doing this would avoid the need for the
AutomationProfile class.

This CL achieves that.
Bug=27695,27662


Review URL: http://codereview.chromium.org/385117

git-svn-id: http://src.chromium.org/svn/trunk/src@32127 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e9e4adcf5b51c08d6ad87035bc6ede199cd7a7b5