linux: use /proc/self/exe when exec'ing ourselves
We don't want to ever go out to disk when looking for data after
startup, since they can be changed by an update.

We *should* just be using the zygote, but the zygote is sandboxed-only
for now, and fixing that has a lengthy dependency graph.  In the
interim, /proc/self/exe is the correct executable and exec'ing it should
work as long as the subprocess doesn't need any other files from the
Chrome directory.

BUG=22703
TEST=Start Chrome.  Move away the entire Chrome directory; plugins
should still work.

Review URL: http://codereview.chromium.org/403018

git-svn-id: http://src.chromium.org/svn/trunk/src@32410 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 19840cc896abb14db3ab89553a50dfe35fda3d0e