Remove the unused ProcessFilter destructor.
Adding some code elsewhere was causing pure call errors related to this
code, but only for Release builds on gcc 4.2.4.  My gcc 4.4.1 computer
was fine.  I know better than to claim it's a compiler bug, but this
destructor is unused anyway so this change is harmless.

Review URL: http://codereview.chromium.org/402080

git-svn-id: http://src.chromium.org/svn/trunk/src@32483 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 9908506c4329aafb19dff1a7493499b8a7d57f2a