Refactors RenderWidget to extract a PaintAggregator class.
After this change, I plan on changing the PaintAggregator algorithm.

Some things to note:

1- Previously, it was possible to send overlapping ViewHostMsg_PaintRect and
ViewHostMsg_ScrollRect messages.  This happened when scrolling a page since the
scrollbar would need to be repainted while the contents of the page are being
scrolled.  With this CL, this overlapping behavior is a bit more explicit.

2- There was a TODO about view_size clipping that I've eliminated.  I was able
to eliminate it because I realized that it is correct to clip the rects passed
by didInvalidateRect and didScrollRect to the size of the RenderWidget.
Apparently WebKit can sometimes invalidate regions outside the view.

R=brettw
BUG=25905
TEST=none

Review URL: http://codereview.chromium.org/403005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@32496 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: bad7e35a6fd7bdd2be0e2dd8232a3c1bf6dbc016