Implement a NTP resource cache.  Instead of re-generating the HTML for the NTP each time, keep a cached copy as part of the profile.
- This changes the handling of StartDataRequest to always happen
on the UI thread since it needs to use the Profile.  Previously
we would generate on the UI thread then handle all the requests
on the IO thread.
- first_view_ got changed into a static bool.  I also removed it
from being added to the json because it didn't appear to be used
anywhere.
- Merge the incognito and regular handler into one since I added
that as a flag to the request in a previous change.

I plan to add the CSS to NTPResourceCache next.

BUG=23120

Review URL: http://codereview.chromium.org/399055

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@32520 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 96c6338a72d7bb7d7ca98290f7f9c2f98aa94fea