NTP: Fix issue where a thumbnail was stuck in hover mode after a delete.
The reason why the thumbnail was stuck is that it was focused due to
the mousedown (the element was reused). Now we blur the element when we
blacklist the thumbnail.

This also adds support for Del (and Cmd+Backspace for mac) to blacklist the
focused thumbnail.

This also fixes the issue where the user could tab to a filler.

BUG=28185
TEST=Remove a thumbnail and move the mouse out of the mostvisited thumbnails.
None of them should be highlighted. 

Review URL: http://codereview.chromium.org/414011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@32573 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ba27b9c19b0e5491e44a271645282ffa3c14efd5