Don't do work in the SharedIOBuffer constructor; use an Init method instead. This eliminates instances of allocated-but-invalid SharedIOBuffers. Also clean up some CHECKs and TODOs associated with the bug.
The upshot is that we no longer crash in AsyncResourceHandler::OnResponseCompleted() when we can't allocate shared memory. We now crash (properly, I believe) in the renderer process if the shared memory that failed to allocate was the TransportDIB, since the renderer can't communicate with the browser without it.

BUG=16371
TEST=none 
Review URL: http://codereview.chromium.org/391009

git-svn-id: http://src.chromium.org/svn/trunk/src@32701 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f56ef24b5c611d0f0d0e68a9f81a754fb983c8f0