Tommi, please review everything. John please review the changes to plugin_service.cc/.h
The test automation provider registers itself as a protocol factory for http/https requests. This 
is to ensure that intercepts set by the url request network tests work correctly. I was seeing these
tests fail consistently on my setup as their intercept function would never get called.

The other change is to add a simple mechanism based on a boolean flag to disable browser side plugins
like gears which also intercept network requests and expect to be called on the IO thread. The chrome 
frame network tests run in a relatively simple environment where the network tests run in a separate thread
(not the IO thread) which causes a number of DCHECKS to fire in debug build test runs. The flag used to
determine whether browser plugins are loaded defaults to true.

Review URL: http://codereview.chromium.org/414017

git-svn-id: http://src.chromium.org/svn/trunk/src@32739 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d565bdbec22999ad993c0d9d771d7c529d741f73