Tommi, please review everything. John please review the changes to plugin_service.cc/.h
The test automation provider registers itself as a protocol factory for http/https requests. This 
is to ensure that intercepts set by the url request network tests work correctly. I was seeing these
tests fail consistently on my setup as their intercept function would never get called.

The other change is to add a simple mechanism based on a boolean flag to disable browser side plugins
like gears which also intercept network requests and expect to be called on the IO thread. The chrome 
frame network tests run in a relatively simple environment where the network tests run in a separate thread
(not the IO thread) which causes a number of DCHECKS to fire in debug build test runs. The flag used to
determine whether browser plugins are loaded defaults to true.

Review URL: http://codereview.chromium.org/414017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@32739 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2c472973fbba530b589d27267f5c2b08969791f1