Mac: give visual feedback for bookmark button drags.
Draw a "ugly black bar" (where "black" is actually a theme colour) to indicate
where a bookmark button that's being dragged will end up if dropped. This is a
stopgap measure for the beta. Later, we will make the buttons move around
instead (even if implementable in the time remaining, such animations would be a
crash risk).

BUG=17608
TEST=Drag bookmark bar buttons around. Make sure that the bar indicating where a button will be dropped is visible enough and accurate. Repeat with various themes.

Review URL: http://codereview.chromium.org/437051

git-svn-id: http://src.chromium.org/svn/trunk/src@33079 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d1d44aab2acf705131236d8e864abcbf8bdd6507