Sync ParseFTPList with Mozilla upstream.
Also fix chromium.patch so that it's possible to apply it.
It turns out my editor removed whitespace from ParseFTPList.cpp anyway,
and somehow the patch itself was corrupted.

The plan is to remove this code after our own LIST parsing code is ready,
but for now we still need it.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/450008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@33312 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 403d61666ddd487522371b05a3dddf3e34e0b283