Don't reuse the initial IPC channel.
I was seeing a non-initial IPC channel getting closed in a renderer, and then
when someone tried to reuse that channel by name,
IPC::Channel::ChannelImpl::CreatePipe would assign the initial pipe.  The
initial pipe was already in use, and things would fall apart pretty rapidly.

I'm making this FATAL because the renderer's probably going to be unusable if
it gets into this state anyway, and a sad tab is probably more useful than a
tab that appears to be loading indefinitely.

BUG=26754
TEST=Test case from bug 26754 comment 9 (affected machines only):
       a. Have lots of bookmarks (import Safari defaults)
       b. Rightclick on bookmark bar, and choose "Open All Bookmarks"
      Expect: no crash.
Review URL: http://codereview.chromium.org/452021

git-svn-id: http://src.chromium.org/svn/trunk/src@33351 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 66463fe61ba954b97af200520463eb32e698e01a