Revert 33344 - Relocate plugin list fetching to PluginService
We fetch the plugin list from three places.  Previously, each location had
custom code to proxy the query to the file thread.  This change moves
the query to the PluginService, which then internally manages posting to
the file thread and calling back.

I experimented with some different approaches to handling the lifetimes
of the requests and responses.  The approach now is simple:
 The PluginService plugin methods are called and respond on the IO
  thread.  Two of the three consumers of the plugin lists are already on
  the IO thread, so they don't need any complicated thread handling.
 None of the callers ever need to cancel their requests: one is a
  singleton, and the other two always wait (in terms of holding a ref
  on the object) for the requests to complete.  This makes lifetime
  management a lot simpler than it would otherwise be.

With this change in place, I can then look at refactoring the PluginService
implementation on Linux to do more complicated plugin loading as needed in
bug 17863.

BUG=17863

Review URL: http://codereview.chromium.org/437069

TBR=evan@chromium.org
Review URL: http://codereview.chromium.org/456012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@33369 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9f62bd3f3756c84e47df771c1d43e6213ae423d6