GTK: add middle click and right click functionality to maximize button.
If the user right clicks on the maximize button, it will horizontally maximize the window, and if the user middle clicks on the maximize button it will vertically maximize the window.

At least on kde, the window manager supports more complicated behavior, e.g. right clicking on a window that is horizontally maximized will return it to its former size. It seems to keep track of multiple restored rects. However, we're not well equipped to replicate this behavior, so I just didn't add it. If the user wants it, they can disable the custom frame.

BUG=28881

Review URL: http://codereview.chromium.org/455012

git-svn-id: http://src.chromium.org/svn/trunk/src@33389 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c8690821d7f090152c2e04ffa631b90878070214