Speculative fix for ChromeFrame crash in bug http://code.google.com/p/chromium/issues/detail?id=29025
The crash occurs while dereferencing the automation channel to send out the SetCookie IPC message on the
automation channel to the host browser. Based on what I could see from the crash dump and the code
it seems like there could be a scenario where the AutomationResourceContext object could be destroyed
while the AutomationCookieStore object is still around and thus ends up with a stale pointer which crashes
when dereferenced.

Fix is to ensure that all related code paths hold on to a refcounted AutomationResourceContext instance.

I will look into whether it is possible to come up with a unit test for this.

Bug=29025

Review URL: http://codereview.chromium.org/450020

git-svn-id: http://src.chromium.org/svn/trunk/src@33524 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 9da99af89125eda8bf232f4f71dfef934c2153f5