Clean up the statistics reported by run_webkit_tests to be easier to follow.
The most important changes are that we bin the failures into mutually
exclusive buckets, so that they aren't double-reported like before. As a
result of this, compare_failures became much simpler and I ended up folding
it into run_webkit_tests and test_expectations

In addition, a lot of the data structures for holding different sets of
failures have been revised to be more consistent and extensible. This allows
me to have the output match the LTTF dashboard as well.

  R=ojan@chromium.org,tc@chromium.org,pam@chromium.org
  BUG=none
  TEST=none

Review URL: http://codereview.chromium.org/414066

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@33640 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5de0e24db20e7cdd454663312dba78342ccda8c2