Don't set referrers on outgoing plugin requests if the load_manually flag is set. This emulates the behavior of other browsers and fixes http://code.google.com/p/chromium/issues/detail?id=28800
I added a UI test to validate that the plugin source URL is set on outgoing GetURL requests issued
by the plugin. To validate that the document URL is set as the referrer on the initial URL request
would take some more work. Will try and add that in a future CL.

I also changed the WebPluginImpl::RouteToFrame function to set the referrer on similar lines.

Bug=28800
Test=Covered by UI test

Review URL: http://codereview.chromium.org/459003

git-svn-id: http://src.chromium.org/svn/trunk/src@33673 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 47a642d1449ca1ab3e71dfc71c975356b68cdbe7