Fixes a bug in the docs generation for overloaded functions.
The issue is that JavaScript has no concept of
overloaded functions. Nixysa was merging them
for compilation by the JSCompiler but it needs
to keep them separate for docs.

To fix the issue an option was added to nixysa
to rename overloaded functions with the suffix
xxxOVERLOAD<num>xxx. The docs generators
then remove the suffix where appropriate.

Review URL: http://codereview.chromium.org/462039

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@33859 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2f7988f2ef1495cd37058703acf6a9cccc323500