Add obscured / unobscured rendering to manipulators.
I only did it on the rotate1 manipulators
because it was easy. On the translate ones
you either need to put each shape on it's own
transform so you can set the parameters different
or you need the materials to use different named
parameters.

Review URL: http://codereview.chromium.org/463025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@33900 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: be7fb4a134152a20c137f42f10d1c7d78ca79b03