Add obscured / unobscured rendering to manipulators.
I only did it on the rotate1 manipulators
because it was easy. On the translate ones
you either need to put each shape on it's own
transform so you can set the parameters different
or you need the materials to use different named
parameters.

Review URL: http://codereview.chromium.org/463025

git-svn-id: http://src.chromium.org/svn/trunk/src@33900 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 429c8e268cefdfc5b1262b165e6e000e1ca65e90