Addition of a new parameter to the popup.show(...) Chrome extension API that allows the caller to specify the behaviour of the window focus when the pop-up is displayed.
I added a test for the new parameter in the ExtensionApiTest.FLAKY_Popup test.  I also corrected a truncated string in the unit-test.

Because the pop-up actually uses two windows, I had to change the WidgetWin::Show() routine to also reposition the window at the top.  If a window is shown and activated, it is automatically brought to the front.  Because we don't activate the pop-up when we want to keep the focus in the current extension view,  I had to bring it to the front so that it wouldn't be hidden behind the 'chrome-bubble' window.

BUG=none
TEST=ExtensionApiTest.FLAKY_Popup

Review URL: http://codereview.chromium.org/454019

git-svn-id: http://src.chromium.org/svn/trunk/src@34037 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3a5c89200412eb34b728d5247b5fbec5413b44bc