Original patch by cbentzel@google.com Original code review: http://codereview.chromium.org/428004
Command line utility to run the HostResolver. 

The user provides a list of hosts to resolve, as well as optional timestamps
for when the resolutions should happen. These can be provided on the command
line as additional arguments or in a specified file.

Options:
--async: Specifies that the resolution should happen asynchronously.
--verbose: Spew out extra data.
--cachesize: Size of the DNS cache.
--cachettl: TTL of DNS cache entries in milliseconds.
--inputpath: File containing host name and optional timestamp, one per line.

BUG=NONE
TEST=Built and ran on Linux and Windows. Built on OSX via trybots but not run.
R=eroman

Review URL: http://codereview.chromium.org/485006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@34225 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f862615b043a0b258df4ea440de9a1a4a8bf49ef