Cache failed DNS resolutions for 1 second.
This is a very small time to live, since we want to be able to respond quickly when formerly unresolvable names become resolvable.

Even such a small time is still useful, since cache misses for unresolvable names can be extremely costly (order of several seconds).

For example, in our corp PAC script, the URL's host is resolved 3 times, so:
  Without caching, total runtime is (2.5 seconds) * 3 --> 7.5 seconds.
  Whereas with caching it would be: (2.5 seconds) * 1 --> 2.5 seconds

This time to live will need to be tuned as part of bug 25472.

BUG=11079

Review URL: http://codereview.chromium.org/464084

git-svn-id: http://src.chromium.org/svn/trunk/src@34238 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 26a554bc85568397db242d4d8b3655aae8eacb53