Cache failed DNS resolutions for 1 second.
This is a very small time to live, since we want to be able to respond quickly when formerly unresolvable names become resolvable.

Even such a small time is still useful, since cache misses for unresolvable names can be extremely costly (order of several seconds).

For example, in our corp PAC script, the URL's host is resolved 3 times, so:
  Without caching, total runtime is (2.5 seconds) * 3 --> 7.5 seconds.
  Whereas with caching it would be: (2.5 seconds) * 1 --> 2.5 seconds

This time to live will need to be tuned as part of bug 25472.

BUG=11079

Review URL: http://codereview.chromium.org/464084

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@34238 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8beac17600608bb7998699607d510c42677049db