The ChromeFrame net tests randomly crash in the url request redirect tests on the chrome frame builder. Some debugging revealed that we send over two redirect responses from the IE host network stack implementation to Chrome, which causes a crash in the url request automation job while dereferencing a NULL request.
Two redirect responses are sent in the following scenario:-
1. We received a redirect notification in our bind status callback. We abort the binding and return E_ABORT.
2. Eventually we receive a call in our bind status callback implementation of OnResponse even after the binding
   was aborted. This causes the response to be sent twice.

Added a check for a NULL binding and a trace in the IE host network stack implementation. I also added
a NOTREACHED in the url request automation job if we ever receive a message for an automation job which
has a NULL request.

Bug=30118

Review URL: http://codereview.chromium.org/487028

git-svn-id: http://src.chromium.org/svn/trunk/src@34360 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 42be628635dc37b2c7fcefefda8e526aa761901a