Fix radio items in views::Menu2 for Linux.
Suppose there are three radio items A, B, C, and A is now being
checked. If you click C, commands associated with A and C are both
executed. This is because A also receives the "activate" signal,
before C receives it. We want only the command associated with A
to get executed.

BUG=30310
TEST=try; manually with view_examples

Review URL: http://codereview.chromium.org/491076

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@34644 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ec6930bc9d3e2e3a4003f337ac0d528740aac67d