Fix browser crash in extension system.
When an external extension is removed from the json file
(for example by a 3rd party installer) after the user
has explicitly uninstalled it, Chrome would start crashing
and crashing.

This is now fixed. Also added a couple of unit tests for
this case (for each external provider) and cleaned up one
of those unit tests to make it match what the other was
doing.

BUG=http://crbug.com/30505
TEST=Covered by unit test.

Review URL: http://codereview.chromium.org/504016

git-svn-id: http://src.chromium.org/svn/trunk/src@34649 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f4dba72dd7e0f11075987db2621f4725d34a1b96