Step 1 in Adding GLES2 renderer to O3D.
This step does only copies the GL files to
a GLES2 folder. It's still actually the GL renderer
at this point but it's the smallest step for 
getting something that builds and a place to
start working.

The next step will be to rename all the 
classes from xxxGL to xxxGLES2

At some point CG will be removed as well

One other thing slipped in there. A standalone
build path to help generate a C++ only example
of using O3D. It's currently only a placeholder.

Review URL: http://codereview.chromium.org/500070

git-svn-id: http://src.chromium.org/svn/trunk/src@34744 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b3649be5ec6fb8ac27fbe048eef3db1b780db59e