Fix some expectations for tests in the chrome and pending directories:
Removed chrome/fast/dom/messageport-gc.html . This was marked as WONTFIX,
and there's not a lot of point in having WONTFIX tests in chrome/, since
these are our tests.

Removed chrome/http/tests/plugins/get-file-url.html . This was marked as
WONTFIX SKIP; apparently it works locally but not on the bots. From looking
at the test, it wasn't very interesting so I've removed it completely
(it's largely covered by chrome/plugins/get-file-url.html ).

Removed pending/fast/repaint/bugzilla-6473 ; this was a variant on the
upstream test that has also been failing for a while, and the only difference
between it and the upstream test was that a patch that was rejected upstream.
Better to fix the upstream test (and we can safely remove this without losing
any history). This is the last file in the pending directory, so we can
remove the entire pending directory.

Also, check in correct mac baselines for chrome/fast/text/fake-italic.html
(they were missing).

  BUG=24195
  BUG=15125
  BUG=30503
  TEST=chrome/plugins/get-file-url.html
  TEST=chrome/fast/dom/messageport
  TEST=pending/fast/repaint/bugzilla-6473.html
  R=pam@chromium.org

Review URL: http://codereview.chromium.org/506025

git-svn-id: http://src.chromium.org/svn/trunk/src@34789 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: fbaca608c507c8017c6b46bd0e54dab93c5f605d