Return a list of changed from WebDatabase::RemoveFormElementsAddedBetween() This changes add a list of AutofillChange out parameter to the RemoveFormElementsAddedBetween() method in order to communicate exactly what happened to the caller.  This method may only update some autofill entries (by removing use timestamps that fall in the specified time range) or completely remove the autofill entry (if all the use timestamps fall between the specified time range).  For sync, we need to know the difference.
This change also required some new testing method on WebDatabase to make it possible to add new autofill entries with specific timestamps.  This is needed to create a reliable test that requires entries to be added at different times.  If there is a better way to do this, please let me know.

The next change will add the notifications to the WebDataService.

BUG=30169

Review URL: http://codereview.chromium.org/506047

git-svn-id: http://src.chromium.org/svn/trunk/src@34889 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c62b3272e0908c45293f57496d670abb0e0a5276