Multiple chrome frame activex controls should instantiate and navigate correctly in IE. This was not the case due to a race condition between put_src getting called for subsequent activex instances and the external tab to hold the chrome frame instance getting created.
Fix is to pass in the URL if we have it when the automation client is initialized to launch the chrome
automation server. If not we navigate when the external tab is created. To achieve this we stuff in
all relevant parameters into a structure which is populated when the automation client is initialized.

I also changed the CreateExternalTab message to carry the referrer for the initial navigation.

Fixes http://code.google.com/p/chromium/issues/detail?id=28236
Test=added unit tests for the same. The firefox one is not working at this point. Disabled this test
     for now while I debug it.

Bug=28236

Review URL: http://codereview.chromium.org/500123

git-svn-id: http://src.chromium.org/svn/trunk/src@34985 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1af8747895c51ec03d3ed22962af65098032a9f7