Mac: Fourth half of CL 481012 (reverted at r35038).
This introduces code to properly enable DnD to the Omnibox without adding a
trivial |-updateDragTypeRegistration| to |AutocompleteTextFieldEditor| (so that
it falls through to the text field), but rather letting the field editor accept
DnD directly.

We'll see whether or not it turns Mac Perf(2) red (by regressing startup times).

BUG=24631
TEST=Select a URL/link/file from somewhere (a link in a browser, a URL in text, a file from the desktop, etc.) and drag it to the Omnibox in a Chromium browser window; the contents of the Omnibox should be selected to indicate that a drop would replace its contents; dropping should navigate to the appropriate location.
TBR=shess@chromium.org

Review URL: http://codereview.chromium.org/500157

git-svn-id: http://src.chromium.org/svn/trunk/src@35065 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1dd5a8f377bb843c33c44fc852eba14bc212e6a1