Revert 35328 - AppCacheDatabase and SQL based AppCacheStorageImpl.
Still nothing is being written to disk with this CL,
inmemory SQLite and DiskCaches are being utilized.
Responses are not yet being removed from the DiskCasche
when the should be. Once that's done (in the next CL), we'll
start saving things on disk.

BUG=none
TEST=appcache_database_unittest.cc, appcache_storage_impl_unittest.cc

Review URL: http://codereview.chromium.org/501033

TBR=michaeln@chromium.org
Review URL: http://codereview.chromium.org/519018

git-svn-id: http://src.chromium.org/svn/trunk/src@35329 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4d88b6236a860533f1eedc69e2b1259b9e5143aa