Fix Popup Blockers close button
The change at http://codereview.chromium.org/484006 changed all the close buttons to use a single HoverCloseButton class, but this didn't work for the popup blocker.  The reason is that all the initialization was done in awakeFromNib since in most other cases, the button was created in a nib, but was done with initWithFrame for the popup blocker case.  I refactored the initialization code out into -initButton to be called from both initWithFrame and awakeFromNib.

I also removed the references to the images and alt images from each of the xib files.  The images are specified in the class itself so it was redundant to have it in the xibs, and could probably cause confusion if we change the images later on.

BUG=16041
TEST=Ensure that the popup blocker button has rollover image changing, and that it works correctly still for TabView, InfoBar, and DownloadShelf

Review URL: http://codereview.chromium.org/521010

git-svn-id: http://src.chromium.org/svn/trunk/src@35568 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 87708e6c111ce98b9134a89e41b67c9401f72ffb