First pass at expanded i18n doc.
We still need to update the overview and perhaps update all
the doc to use/reflect i18n. And we need to add a page for
the messages file format. (For now, we just give an example
and point to the design doc.)

Only 2 files really changed: i18n and manifest.
All the other changes are due to adding i18n to the sidenav.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/525077

git-svn-id: http://src.chromium.org/svn/trunk/src@35661 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1b24bb02f69f2684f1d853bf1b05c2b7b9d896ad