linux: make the seccomp sandbox work again
We were hitting a stack overflow on renderer startup, because of the following:
When we patch out syscalls, we need a scratch space near (within a 32-bit jump)
of the original code.  We pick the scratch space as the end of the nearest empty
region available before the code we're patching.  For the vdso region, the
stack lies directly before it and so the region we'd grab was directly before
the stack.  This meant that as soon as the stack attempted to grow it'd fail
because it ran into our patch region, and we'd hit a stack overflow.

The fix is to specially note when we're near the stack region, and instead put
our scratch space as far away from the stack as possible.

Review URL: http://codereview.chromium.org/518071

git-svn-id: http://src.chromium.org/svn/trunk/src@35759 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 1afdaeb2130d12f4435a66553ff32a489898cbdc