Adds a NetworkChangeNotifier dependency to ProxyService; when we observer a network change through the notifier, the ProxyService re-configures itself.
So for example, if you were to switch to VPN and are using a custom PAC URL, we will re-download that URL on the new network.
Similarly if you are using auto-detect PAC settings, we will re-run the autodiscovery steps for the next resolve.

Note that Chromium is still passing a NULL NetworkChangeNotifier dependency into its ProxyService -- until it passes a real implementation, this will not impact Chrome.

BUG=12293
TEST=ProxyServiceTest.NetworkChangeTriggersPacRefetch
Review URL: http://codereview.chromium.org/525104

git-svn-id: http://src.chromium.org/svn/trunk/src@35772 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: c977b588a41b2fcbfb4b1c3375c2fe2f24ec413d