Take 2: Preload the visited link db on the file thread if the file exists. Otherwise, just load like normal on the UI thread.
This failed before because the browser may have shutdown before the
posted task ran.  When the posted task finally runs, it tried to
use the profile, but the profile was already deleted.  Make a small
change to VisitedLinkMaster so GetDatabaseFileName no longer depends
on the profile.

BUG=24163

Review URL: http://codereview.chromium.org/507047

git-svn-id: http://src.chromium.org/svn/trunk/src@35991 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0ad23654a453548e41f572d367ca9b3074893e81