SSL: don't ask for next proto if we got an error.
GetNextProto only returns an error if the NSPR socket passed in is not a valid
SSL socket. However, this happens in tests (see the bug).

It appears that in some code paths the socket is invalid by the time that we
call HttpNetworkTransaction::DoSSLConnectComplete.

This also fixes the case where we would use SPDY if the server supports NPN,
but didn't actually advertise support for it.

BUG=31611

git-svn-id: http://src.chromium.org/svn/trunk/src@36024 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0973817301aaef25277d6085547600bddc61d0b5