Change bookmark manager object model.
This is in preparation for adding Recents/Search. Apologies for the number of diffs, but most of them are basically just renaming, especially 'id' --> 'BookmarkItem*' and 'node'-->'item'.

The code now has a real Objective-C BookmarkItem class instead of using opaque 'id's. This allows all the C++ BookmarkNode calls to be collected into one source file. It will also in the future allow creating fake collections of bookmarks, necessary for the Recents and Search features.

Also some refactoring of the drag/clipboard code, which now uses a static array instead of an instance variable to hold dragged items; this will be necessary to handle moves between trees (coming in the next checkin.)

Minor xib changes: Shorten URLs in the middle not at the end, and remove borders of table/outline views.

BUG=32080
TEST=Added BookmarkItemTest, modified existing unit tests.

Review URL: http://codereview.chromium.org/551013

git-svn-id: http://src.chromium.org/svn/trunk/src@36195 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 51bf79d6a3335dba87a44d485cdc8a8aa931cc84