Rename CloseIdleConnections -> CloseCurrentConnections.
This method is (and was) just for debugging; while it was closing
idle connections, that was not aggressive enough; the benchmark needs
to close all connections, not just the idle ones.

To ensure connections are abandoned, create a new pool; leaving the old pool
to languish as any pending sockets die.

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/549031

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@36208 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 34d39fb4bbbf67dcf505c9ea4487b33ed5bb2b17