Fix a Chrome crash which occurs when the chrome instance launched via ChromeFrame is shutting down. I was not able to reproduce this locally. Based on the crash dump, it looks like the crash occurs when the sync channel used by the automation provider is shutting down and it is attempting a PostTask of OnChannelClosed. The crash occurs while posting the task to the message loop, because the message loop is invalid.
This could occur if the message loop has been destroyed. The message loop is destroyed when the module
refcount drops to zero. The automation provider releases the module ref count in OnChannelError. In theory
there is a race condition here, i.e. the message loop could be destroyed before the ExternalTabContainer,
etc are destroyed.

Suggested fix is to call AddRefModule which grabs a reference on the UI message loop in the automation provider
constructor and release it in the destructor. We also close the channel before invoking ReleaseModule in the
destructor.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=31621

Bug=31621

Review URL: http://codereview.chromium.org/542036

git-svn-id: http://src.chromium.org/svn/trunk/src@36240 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7e9fc4f18472db382641cb6ccb09bfb3182b5a84