DevTools: fix interactive tests flakiness appeared after http://src.chromium.org/viewvc/chrome?view=rev&revision=36518.
The console messages are pushed to the frontend on attach as it used to be before. To ensure that injected script source is available when WebDevToolsAgent::attach is called it is loaded from Chromium resource(temporary measure until we come up with something better).

BUG=29068

Review URL: http://codereview.chromium.org/546086

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@36635 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c6e960a612418c59b8b2b006792be3a18db8affd