Fix js alert crasher when window is closed by js
There was a problem with pages that call window.close() then immediately open a js alert.  The problem was that the TabContents would get destroyed, but the alert would stick around, and when you closed the alert, it would try to call back to the tab contents' js, crashing chrome.  The other OSs got around this by closing the alert immediately, but mac didn't implement the CloseModalDialog method.  This CL implements that.

BUG=29580
TEST=Run the html attached to the bug

Review URL: http://codereview.chromium.org/549126

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@36908 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c9b2c956c62cf46fbede3c6f12ec109f2c538233