Extension tweaks for phantom tabs. I did the following modifications:
. Modified chrome.tabs.executeScript so that it sends error message if
  failure in sending message (which happens with phantom tabs).
. When a tab is made phantom we send TabInsertedAt event.
. Made connecting to a phantom tab send disconnect.
. Disallow changing the url of pinned tabs.
. Disallow closing phantom tabs.
. Detect language fails for phantom tabs.

And I removed the flag for phantom tabs.

BUG=25309
TEST=on windows or chromeos create a window with a couple of tabs, pin
the first, then close it. Selection should change to another tab and
the favicon of the closed tab should remain, but not the tab
border. control-tab should not select the phantom tab, but if you
click on the phantom tab it should become live again and the border
should reappear.

Review URL: http://codereview.chromium.org/552110

git-svn-id: http://src.chromium.org/svn/trunk/src@36964 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 28e2abecf7875f00cfaf4ee43b706f9cb72fe5a4