Better handle oversized IPC messages
* Shoot down oversized messages on the sending side so we fail faster.
* Add DCHECKs to identify oversized messages early.

The real fix for the underlying bug is not to send oversized messages in the first place, but the current state of things is that it takes a long while for the renderer to crash.  This change should speed the failure up a bit.

BUG=26822
TEST=Chrome should continue to load web pages.

Review URL: http://codereview.chromium.org/546047

git-svn-id: http://src.chromium.org/svn/trunk/src@37102 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 620071ae269e7323d8316f3bad689646e604f2b4