Update webkit to r53957.
This includes changes to WebKit that make history.{back,forward,go}
behave synchronously when navigating back to a reference fragment.
That is something we still implement asynchronously.  As a result,
we now fail the location-hash.html test.

Includes changes to glue_serialize.cc to support serializing the
WebHistoryItem's documentSequenceNumber.  (+sky for review on this
part.)

TBR=pfeldman,sky

Review URL: http://codereview.chromium.org/556055

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@37388 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7d65a6f2ce51f8a3f0a2795b5c23bb2b3c6e1322