Update webkit to r53957.
This includes changes to WebKit that make history.{back,forward,go}
behave synchronously when navigating back to a reference fragment.
That is something we still implement asynchronously.  As a result,
we now fail the location-hash.html test.

Includes changes to glue_serialize.cc to support serializing the
WebHistoryItem's documentSequenceNumber.  (+sky for review on this
part.)

TBR=pfeldman,sky

Review URL: http://codereview.chromium.org/556055

git-svn-id: http://src.chromium.org/svn/trunk/src@37388 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f5630f876378889152b2aebd2cd628d8cb09b2a5