Resolve issues with GlueSerializeTest.BackwardsCompatibleTest by appending the new documentSequenceNumber entry to the end of the pickle.
This test forges kVersion, but the function to serialize a
WebHistoryItem does not take the current value of kVersion
into account.  It did not do so before, so instead of trying
to make it do so now, I just chose to append the new field.

TBR=sky

Review URL: http://codereview.chromium.org/556057

git-svn-id: http://src.chromium.org/svn/trunk/src@37391 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: b04adcce5e4a44adff4b8fa7c1c6f4ad22fc9af4