Add commands Jump, Call and Return
This is the minimal change version.

There were 3 possible solution:

#1) Remove these commands

#2) Change the command buffers so the client side
can not see the "get" pointer and therefore force
the client side to use tokens to say in sync.

It's the cleaner solution. It lets you put commands
in any shared memory thereby getting rid of the ring
buffer commands. 

But, it required major changes. After going down
that path for an entire day I punted for...

#3) Implement the commands assuming there is only
one buffer the commands can be in (what is currently
called the ring_buffer). With the jump commands
available it no longer has to be used as a ring
buffer. A Pepper or NaCl app can use it however
they want but, unlike solution #2 they only get
the one buffer.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/555129

git-svn-id: http://src.chromium.org/svn/trunk/src@37446 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 4d7d5178a84279538a53668ba2b32ac72ab28301