In IE8 new windows opened within ChromeFrame via window.open calls at times bypass the host network stack.
In this case we don't get control over the navigation as window.open calls expect to carry the opener
relationship over to the new window. This basically means that navigation occurs in Chrome and the 
new tab/window created by the host attaches to the newly created ExternalTabContainer instance.

In IE8 the new tab opens in a new IE process, which basically uses a new automation channel to talk to
Chrome. The host network stack implementation routes network requests issued by registered render views
to the host browser over the automation channel. There is a timing window between the new window getting
created and issuing network requests and the channel being established with the new iexplore instance.
As a result network requests issued by the new window don't use the host network stack.

Fix is to register the render view and process as a pending view when we get notified about the new TabContents
in the original ExternalTabContainers implementation of TabContentsDelegate::AddNewContents. Any network requests
issued for this view would result in the corresponding URLRequestAutomationJob instances getting created as 
pending as well. When the host browser connects to the new ExternalTabContainer instance, we pass over the new
automation channel and tab handle to the URLRequestAutomationJob instances and resume them.

This fixes bug http://code.google.com/p/chromium/issues/detail?id=33516

Bug=33516
Test=Login to gmail in IE8 in ChromeFrame. Open up a new conversation and click on New Window which opens up
     a tear off window. This should not bring up the login prompt again.

Will think about a good approach to write a unit test for this behavior and send out a separate CL for that.

Review URL: http://codereview.chromium.org/554134

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@37585 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 714094d6dc3b2604f156ed06d0fbf2b32796529f