Modify the toast to hit system-level installs
This is somewhat tricky. setup.exe is executed by SYSTEM and
as such we cannot run the experiment. So what we do is that
if there is an interactive user logged in at the moment of
chrome upgrade we re-launch setup in her context, from then
on things proceed as in the previous case.

Other enhancements
- New experiment codes TKxx and TLxx (99% and 1% as usual)
- New experiment group TK40 and TL40 : An active user

BUG=32474
TEST=see bug

Review URL: http://codereview.chromium.org/557077

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@37799 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 08afab41384941a1739589fbe3cfcbf7a79fdf6c