[Mac] Make the extension popup a singleton instance to avoid maintaining multiple pointer references to a class that, by design, can only have one popup open at a time.
The bug fixed in this change has to do with PageActionView having a dirty pointer to a popup if it was closed via losing its key state. Once a notification like EXTENSION_HOST_VIEW_SHOULD_CLOSE was fired, then the PageActionView would assume the pointer was valid and call on dirty memory.

TEST=none
BUG=29492,33590
Review URL: http://codereview.chromium.org/561013

git-svn-id: http://src.chromium.org/svn/trunk/src@37873 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d155992f1d6fe8d26708ae58f8896f3736c4a32e