Support app menu and double-click close on upper left corner of window, part one.
This makes things work right for windows that have an icon.  The behavior of the main window is unchanged because right now it never returns HTSYSMENU anywhere.  Part two of this change will be figuring out how to return that at the right place, which will make things Just Work.

BUG=6961
TEST=Open an app or popup window.  Single click the icon.  The app menu should appear.  Close it, then click-and-drag into the app menu.  The item you release on should execute.  Double-click the icon.  The window should close.
Review URL: http://codereview.chromium.org/577018

git-svn-id: http://src.chromium.org/svn/trunk/src@38187 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 20c24dfcb5a9b62b3ac8094cbf950d91f1b8dbdb