Fixes crash in making a tab phantom. Here was the sequence when making a tab phantom: 1. close the old contents (sending out TabClosingAt). 2. TabSelectedAt 3. TabReplatedAt
The problem is is Browser::TabSelectedAt processes updates for the old
contents, resulting in:

1. close the old contents (sending out TabClosingAt).
2. TabSelectedAt
a    UpdateTabContentsStateAt (for old tabcontents)
b    TabChangedAt             (for old tabcontents)
3. TabReplatedAt

So, 2b is sent for a tabcontents that observers was already told
closed in step 1. This triggered a DCHECK in extensions code.

I've changed it so we send out TabSelectedAt after the new contents
are swapped in.

I also fixed a bug where I wasn't telling observers the tab strip
empty if the last non-phantom tab is made phantom.

BUG=34137
TEST=none

Review URL: http://codereview.chromium.org/579024

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@38363 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4d44624729825e6228097e0973f8735139bd7e14