Fixes crash in making a tab phantom. Here was the sequence when making a tab phantom: 1. close the old contents (sending out TabClosingAt). 2. TabSelectedAt 3. TabReplatedAt
The problem is is Browser::TabSelectedAt processes updates for the old
contents, resulting in:

1. close the old contents (sending out TabClosingAt).
2. TabSelectedAt
a    UpdateTabContentsStateAt (for old tabcontents)
b    TabChangedAt             (for old tabcontents)
3. TabReplatedAt

So, 2b is sent for a tabcontents that observers was already told
closed in step 1. This triggered a DCHECK in extensions code.

I've changed it so we send out TabSelectedAt after the new contents
are swapped in.

I also fixed a bug where I wasn't telling observers the tab strip
empty if the last non-phantom tab is made phantom.

BUG=34137
TEST=none

Review URL: http://codereview.chromium.org/579024

git-svn-id: http://src.chromium.org/svn/trunk/src@38363 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3f5b60cdf904e3be27dc78ba3b58e6d657b3516b