Make it so that chrome.bookmarks.update can update the URL of a bookmark.
This changes the API slightly. Previously, if someone called update with an empty object ({}) it would set the title to "". This was only documented in the code and not on the API page. Now that we support updating both the url and title it seems more reasonable to ignore a missing title.

BUG=34841
TEST=browser_tests.exe --gtest_filter=ExtensionApiTest.Bookmarks

Review URL: http://codereview.chromium.org/591006

git-svn-id: http://src.chromium.org/svn/trunk/src@38555 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 03936aec1c0698cd4fa14f09ffc48e7b94bb90f3