[Linux]Fix text justification of omnibox.
Without this CL, the omnibox inherits the default behavior of GtkTextView on handling text direction and justification, that is:
1. If the justification is set to left justification, then the content will be left justified if the content's text direction is LTR, otherwise it'll be right justified.
2. If the justification is set to right justification, then the content will be right justified if the content's text direction is LTR, otherwise it'll be left justified.

And if the content doesn't have a strong text direction, the keymap's direction will be used. If the keymap doesn't have a strong text direction either, the widget's direction (UI language direction) will be used.

This behavior is different than Chrome windows. On windows, if the UI language direction is LTR, then the content is always left justified, otherwise it's always right justified.

To achieve the same behavior, this CL adjust the justification automatically according to the content's (or keymap's) text direction and the widget direction, to make sure the justification is always in sync with the UI language direction.

BUG=34724 Text in omnibox and find bar is not right justified correct under RTL locale.
TEST=Run chrome with LTR locale, the text in omnibox should always be left justified, regardless of the text's direction. With RTL locale, the text should always be right justified. It matches the behavior of windows version.

Review URL: http://codereview.chromium.org/595021

git-svn-id: http://src.chromium.org/svn/trunk/src@38737 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a36904407bbddef04ccda6ea20ac13b9c9a15375