Change ViewHostMsg_SetCookie to be routed so we can show the blocked cookie notification.
We get the RenderView identifier by looking at the current WebFrame.
This will need to be changed since we do not always have a current
WebFrame (e.g., when processing a <META http-equiv="set-cookie">).
I'd like to proceed with this temporary fix for now since I think it
is actually good enough for a 4.1 release.  I'll work to make the
necessary webkit changes to expose the WebFrame properly in parallel.

R=jorlow
BUG=34573
TEST=none

Review URL: http://codereview.chromium.org/605003

git-svn-id: http://src.chromium.org/svn/trunk/src@38939 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: f1221f7ab2202b9e4fcd15173fcfcd00868a2484