Reverting change: broke unit tests on chrome OS.
Revert 39143 - This change will include preferences as part of the usual startup sequence of the PSS that already includes bookmarks.  This is a temporary solution until we have a proper component to manage the startup and shutdown of multiple data types.

Review URL: http://codereview.chromium.org/601037

TBR=skrul@chromium.org
Review URL: http://codereview.chromium.org/611006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@39155 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: fb04eb988cb4a42cc4c884dc5bbb50b85971eb62