Modal loops of joy.
We need to pump messages in the renderer and in any related plugin processes
when sending sync IPCs to get cookies and set localstorage items because these
IPCs can block on a user prompt.

This code moves complexity into RenderThread::Send().  That function is now
responsible for calling WebView::{willEnter,didExit}ModalLoop() and for
broadcasting the messages to signal/reset modal dialog events in the related
plugin processes.

This change also adds code to optionally suspend the shared webkit timer.  This
was needed to fix bug 36063.  It also helps make spinning a nested loop to wait
for document.cookie results a bit less risky since we'll be re-entering WebKit
in fewer cases.

R=jam
BUG=34917,36063
TEST=none

Review URL: http://codereview.chromium.org/607011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@39327 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 20ace749d9f4d3abc40d84202c2e58f6df13b921