linux: clean up a bit of redundant logic
I noticed this while looking for something else.  In theory, this
will reduce flashing when the dialog comes up (since we first do
all the layout of the contents and then once that's all done we
show the window), but I haven't verified that.

Review URL: http://codereview.chromium.org/608034

git-svn-id: http://src.chromium.org/svn/trunk/src@39345 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5995450aafb01eab45bff919863e405aaf3dee50