Fix for referrer flakyness.  There were two problems:
Grabbing the referrer header when the request is made and not in BeforeNavigate2.
The request headers aren't always available in BeforeNavigate2.  Mshtml provides these
headers in IHttpNegotiate::BeginningTransaction, which is where we now grab them.

There was a race in chrome_frame_automation.cc where we use std::string to store the referrer url and then access that string from different threads.  This was causing the referrer URL to be missing at some times and at other times be sent over to chrome as bad string (e.g. containing embedded nulls) and subsequently be deemed an invalid URL and dropped.

TEST=Fixes referrer header flakyness.
BUG=34812

Review URL: http://codereview.chromium.org/646013

git-svn-id: http://src.chromium.org/svn/trunk/src@39383 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 9c629dd9585dfa21de7767333a65ac3321554175