Fix a Chrome crash caused in a ChromeFrame instance while displaying the inspector window. The inspector window uses the browser view which instantiates the whole view including the ToolBar, which in turn instantiates the Autocomplete edit bar which uses Richedit on Windows. The auto complete edit bar dynamically loads and unloads the richedit control and it uses ATL to superclass the richedit control. If the rich edit dll loads at a different base address it causes a crash in Chrome while creating the window as ATL caches the wndproc address for a class.
Fix is to setup an implicit dependency on the riched20.dll and not load and free it all the time. To achieve this we use #pragma comment for the 
riched20.lib and a dummy call to the CreateTextServices function exported by the dll to ensure that the linker does not discard the import.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=33308

Bug=33308

Review URL: http://codereview.chromium.org/646043

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@39415 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d0be0fe1f6018f1bc53788403730c6e2f6db9f76