Add "Force Reload This Page" menu item on Mac.
This CL adds a menu item "Force Reload This Page" with keyboard shortcut Command-Shift-R to the View menu in MainMenu.xib as an alternative, meaning it is shown instead of "Reload This Page" when the shift key is held down.

BUG=36245
TEST=Open the view menu. Pressing shift should change "reload this page" into
"force reload this page", which should do what it says. Also, Command-Shift-R
should do the same thing. Astronomy picture of the day is a good test site for
this.

Review URL: http://codereview.chromium.org/651040

git-svn-id: http://src.chromium.org/svn/trunk/src@39594 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 591a866bbac5e9e05b51b560636d250fed129353