WebKit roll 55080:55089
Also merging in this Chromium side patch to match:

Linux: add plumbing for fontconfig on per-strike basis

fontconfig on Linux can change the render preferences on a per strike
basis (a strike a combination of face and size). Because of this, we
need to query fontconfig each time a new FontPlatformData is created
for a new size.

This is the Chromium side of https://bugs.webkit.org/show_bug.cgi?id=33065

BUG=18159
TEST=A new renderer should respect settings in ~/.fonts.conf

git-svn-id: http://src.chromium.org/svn/trunk/src@39616 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a95416df01ed5bb64f91c43fc0471782be452c1f