WebKit roll 55080:55089
Also merging in this Chromium side patch to match:

Linux: add plumbing for fontconfig on per-strike basis

fontconfig on Linux can change the render preferences on a per strike
basis (a strike a combination of face and size). Because of this, we
need to query fontconfig each time a new FontPlatformData is created
for a new size.

This is the Chromium side of https://bugs.webkit.org/show_bug.cgi?id=33065

BUG=18159
TEST=A new renderer should respect settings in ~/.fonts.conf

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@39616 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 8803db3fcb959e02e70678bfd920cdfb517b112e