Temporary fix for issue 34061.
Change that signals that drag-drop operations are immediately terminated in ExtensionHost views.  Previously, the view would become unresponsive, because the RenderViewHost would remain 'stuck' in the dragging logic, and all input events would be disregarded.

This fixes the issue described in the bug, but the long-term solution is to properly implement drag-drop operations for extension-views.

BUG=34061.
TEST=None.

Review URL: http://codereview.chromium.org/650059

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@39627 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 742fff93ab19af54b9c73757d9f990d61f2d1e6e