Fixed a hang in ChromeFrame while tearing down pending url requests during shutdown. This would occur when we received a redirect notification on an aborted request. We would end up aborting the request twice which would cause the UrlmonUrlRequest object to be destroyed twice by IE, thus resulting in a junk reference in the request map.
Fix is to bail out in the IBindStatusCallback::OnProgress callback if it is invoked for an aborted request.

Part of the fix for http://code.google.com/p/chromium/issues/detail?id=34687

The other issue being fixed is a crash in ChromeFrame caused if the UrlmonUrlRequest object gets destroyed
in the context of UrlmonUrlRequest::StartAsyncDownload. The crash would occur while accessing members of the
object on return from this function. Fix is to grab a reference on this object in the context of the call.

Bug=34687

Review URL: http://codereview.chromium.org/652202

git-svn-id: http://src.chromium.org/svn/trunk/src@39909 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: eb251e4db1c976762ed75cea9fccbff753294c3b