Fixes ReadPixel to have a result so the client side can know if it should or should not copy the results.
This is the way the spec is. If ReadPixels fails
it's not supposed to modify client side memory.

Unfortunately that makes it slow because it stalls
the GPU pipeline 3 times.

The service side has to  call glGetError to save up 
any current errors. That's one stall. Then it calls
ReadPixels, another stall, finally calls glGetError 
a 3rd time to see if  ReadPixels succeeded so that 
it can tell the client side implementation of
ReadPixels whether or not to copy memory.

We could decide to change the behavior of ReadPixels
so it always copies. That wouldn't be spec complient
but suspect most developers never expect it to fail
and don't count on it not always writing to their
memory.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/652213

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@39930 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 097cb8823a5ebf6113353fbe456a957ca26e8b3e