Don't re-run content scripts on fragment navigations.
This regression was inadvertently introduced by a fix to make executeScript 
run after a fragment navigation (bug 29541). That patch was:

http://codereview.chromium.org/566041

The problem is that on frame navigations (didChangeLocationWithinPage), we end
up creating a new UserScriptIdleScheduler, so this patch keeps track of the 
previous has_run state and propagates that to the new UserScriptIdleScheduler.


BUG=35924
TEST=Steps to verify are outlined in bug report


Review URL: http://codereview.chromium.org/646017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@39939 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7d3ed3920884631954aafd171a60d6a8ba9d3001