Implement WebCookieJar.
Moves the implementation of cookie methods in RendererWebKitClientImpl
into RendererWebCookieJarImpl.

Moves the implementation of cookie methods in TestShellWebKitInit into
SimpleWebCookieJarImpl.

This leaves us with no longer implementing the cookie methods on
WebKitClient.  That'll allow us to delete those.

WebPluginImpl needs to access the WebCookieJar used by either the
renderer or test_shell.  To enable that, I added a GetCookieJar method
on WebPluginPageDelegate.  This interface is used whenever the
WebPluginImpl needs to talk to the RenderView or TestWebViewDelegate.

R=jorlow
BUG=35589
Review URL: http://codereview.chromium.org/661070

git-svn-id: http://src.chromium.org/svn/trunk/src@40023 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e9644cff2936f5458961d1dd253ff812831f51dd