Cancel bookmark editing when changes happen suddenly.
BUG=http://crbug.com/32896, http://crbug.com/29122,
	   http://crbug.com/34369, http://crbug.com/34513,
	   http://crbug.com/34522

TEST=1)
Add some bookmarks.
Go to one of them.
Open the bookmark manager so that the current item is visible.
Click STAR to open the bubble.
Right-click on the item in the bookmark manager and pick Delete.
Watch bubble go away safely.
2) 
Create a bookmark folder.
Open the bookmark manager so that the folder is visible.
In Browser, right-click on folder in bar and pick Rename (to open sheet).
Now right-click that folder in the bookmark manager to edit it.
  And edit it in there...
On "return", the Edit sheet in the main browser window goes away.
3)
Right click any existing folder on BMB and Choose "Rename" option
(Leave sheet open...)
Open other new window via menu or dock and "Delete" the folder which we tried to "rename"  on first window
See sheet close in 1st window

Review URL: http://codereview.chromium.org/661335

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@40522 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 450debc40006cd9c243929b1fe752e2257272763