Change histogram names because collected data is somewhat different.
Net.HadConnectionType3
Each bucket is a boolean (0 or 1) indicating whether the user has had a successful connection of that type during the session. The bucket indexes are defined in "net/base/connection_type_histograms.h". 0 = Any connection (SSL or not); 1 = An SSL connection; 2 = An SSL connection with an MD5 certificate in the certificate chain (excluding root); 3 = An SSL connection with an MD2 certificate in the certificate chain (excluding root); 4 = An SSL connection with an MD4 certificate in the certificate chain (excluding root); 5 = An SSL connection with an MD5 CA certificate in the certificate chain (excluding root); 6 = An SSL connection with an MD2 CA certificate in the certificate chain (excluding root); 7 = A HTTP connection. 8 = A SPDY connection.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/661409

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@40545 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 71f0da73a25782fbbe8c28b3489353b7cf71f344