A follow-up of r28664.
Remove the unused member completed_handshake_.

In OnHandshakeIOComplete, correct the comments to note
that we also notify the caller of success.  We cannot
DCHECK renegotiating_ is true because DidCompleteRenegotiation
has reset it to false.  Instead, DCHECK the equivalent
conditions in DidCompleteRenegotiation.  Log the end of
TYPE_SSL_CONNECT only for initial handshakes.

In DoCompletedRenegotiation, result may not be OK.

R=mbelshe
BUG=none
TEST=none
Review URL: http://codereview.chromium.org/666004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@40550 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d136262d79458871867b61673e489486ca392d20