Fixed a startup race condition.
Although the new test was written in a platform independent way, it is only added to the Widows specific portion of the ui_test target in the gyp file because it wasn't tried yet on the other platforms.

The bug was found and the fix was written in the windows specific version of the process singleton anyway... But if people working on the other platforms would like to try the test there, that would be great. :-)

BUG=9593
TEST=A new test have been created to validate this.

Review URL: http://codereview.chromium.org/661339

git-svn-id: http://src.chromium.org/svn/trunk/src@40629 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 3b85d0ad55b2d553957a4f0da28bc045384322e2