Switch to continuous reporting of uptime
Uptime was being reported only at startup, and then, for the previous session.
This meant that folks with a long uptime would be very slow to report.
This change moves the reporting to an incremental reporting, done as
often as we report pageloads etc. (i.e., in each UMA upload).

This also fixes several bugs relating to uptime calculation, by using
TimeTicks (which show process time) rather than Time (which shows wall
clock time, and is subject to corruption if the user changes the will
clock time on the computer).  It also resolves a bug that may have caused
negative uptimes to be reported.

BUG=37630
r=hunar
Review URL: http://codereview.chromium.org/675001

git-svn-id: http://src.chromium.org/svn/trunk/src@41216 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 42767e42d363f276ad5796624daab95e98459e38