Mac: Improve window look from "completely broken" to "pretty broken" when RI is enabled.
From http://developer.apple.com/mac/library/DOCUMENTATION/UserExperience/Conceptual/HiDPIOverview/HiDPISupport/HiDPISupport.html:

"Because of scaling, the coordinates of the window frame and its top-level view (the frame view) are not always the same. [...] These methods provide the means to transform geometry to and from a base coordinate space that is pixel-aligned with the backing store into which the view is being drawn."

Stuff is still pretty broken with this patch (it used to be better), but it's a low-hanging fruit.

BUG=19476
TEST=Open QuartzDebug, set scale factor to e.g. 1.5. Reopen Chrome. Tabs should be visible.

Review URL: http://codereview.chromium.org/893001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@41336 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9ccda7d75ecd646efdbdc81d08574a2cb2a0fad4