NSS and NPN: send NextProtocol on resume and disable NPN without SPDY.
NPN on resume was broken symmetrically between OpenSSL and NSS (which
is probably why we didn't notice for so long). OpenSSL wasn't
expecting the message and NSS wasn't sending it.

This change sends the NextProtocol message when resuming and disables
NPN when not using SPDY.

The reason for the latter is that we have GFEs deployed which don't
expect NextProtocol when resuming and we have to wait some time before
they get replaced. Because of this, we can't have NPN enabled by
default. Instead, we'll use --host-resolver-rules to redirect
everything to known-good GFEs in the mean time.

http://codereview.chromium.org/841003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@41433 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 38aaa7c4ab095f773cb348eb921cf223582c43be