[Mac] Better code sharing between cookie tree and cookie prompt
Refactored code so that the class that encapsulates cookies details
is shared between the cookie tree and cookie prompt on the mac.

xib changes: Fixed the bindings in all of the subviews in the details view to use the new shared details class. Also added a currently hidden popup menu for selecting the cookie expiration time in the prompt details view--this needs to be wired up in a separate CL.

BUG=http://crbug.com/38348
TEST=added unit tests for the new class, manual testing of cookie tree and cookie prompt dialog

Review URL: http://codereview.chromium.org/1025003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@41949 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 2938b4f2c8fce557f24234ca9e00491377fc426b