Fix the initial draw handing of Core Animation plugins
Removes the call to SetWindow(NULL) in the Core Animation case, since that causes the delegate to enable the TransportDIB drawing mode. Removes the early return in Paint, since now we don't call it (and if we do later, it'll be to get snapshots using the existing Cocoa event model path, per the ammended CA drawing model spec).

Also renames windowless_ in WebPluginDelegateProxy to make it clear that it's just about drawing.

BUG=37974
TEST=In a build with Core Animation plugin enabled, there should be no flash of black when the plugin is first drawn.

Review URL: http://codereview.chromium.org/982010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@41961 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d9d80e354da3d4aa1e8c0f3106c783402d7deab3