Fix Issue 33145: Content Editable Move Image or Selected Block of Text Broken
The cause was that the original source mask only specified "Copy" as allowed operations, while the destination operation is "Move".

Changed the code to allow both "Copy" and "Move" (in line with the GTK implementation). Note that this patch does not touch or implement the TODO(snej) issues, i.e., allow for operations other than move or copy.

BUG=33145
TEST=see bug description

Review URL: http://codereview.chromium.org/1097001

git-svn-id: http://src.chromium.org/svn/trunk/src@42069 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e89802eccca97be2a127c717c454e02dc0b89b30