Implement SOCKSClientSocketPool
This is the first layered pool, so there are several infrastructure changes in this change as well.
Add a ConnectionTimeout method to pools so that layered pools can timeout each phase.
Add a name method to pools to support per pool UMA histograms.
Change SOCKS sockets to take a ClientSocketHandle instead of a ClientSocket

BUG=30357 (blocks an SSL Pool)
TEST=existing unit tests

Review URL: http://codereview.chromium.org/668097

git-svn-id: http://src.chromium.org/svn/trunk/src@42231 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: ea029e01d4453c6c513b3f157e108fb24befc513