Fix ChromeFrame referrer test which was flaky due to incorrect http header parsing. To pass the referrer to Chrome during navigation we were converting normalized HTTP headers which are terminated with \r\n to raw headers and attempting to parse them. Converting to raw headers basically replaces \r\n with the NULL terminator which causes the parsing to fail as a result of which we fail to retrieve the referrer.
Fix is to not convert the headers to raw and just parse the orignal headers.

The other change is to only flag success or failure in the referrer frame test if we are running in Chrome.
This page loads initially in IE as well due to the vagaries of the http equiv patch. Flagging an error
if we don't find a referrer basically causes the test to fail.

This should fix http://code.google.com/p/chromium/issues/detail?id=34812

Will remove the flaky attribute on the ChromeFrame Referrer test in a subsequent CL if all goes well.

Bug=34812


Review URL: http://codereview.chromium.org/1128013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@42372 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e614069c8ea786aa242cede4d90ed62f672ddbb7