Add IPv6 probe support for Windows
This is meant to test to see if AI_ADDRCONFIG is doing its
job in windows, or if we have to do the scans when there is
a network change.

Also switch to always doing probe support on Mac/Linux platforms.
There appears to be (on aggregate) around a 40% degradation of
DNS resolution times on the Mac, and about 17% degration on
Linux, when we don't probe.  It is likely that a few users
are greatly impacted by this, and are able to move the average
of the whole user population.


BUG=12754
BUG=25680
r=wtc
Review URL: http://codereview.chromium.org/1210003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@42544 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9492547cb35711c1c72cc418030565e8ac1abf1e