Add IPv6 probe support for Windows
This is meant to test to see if AI_ADDRCONFIG is doing its
job in windows, or if we have to do the scans when there is
a network change.

Also switch to always doing probe support on Mac/Linux platforms.
There appears to be (on aggregate) around a 40% degradation of
DNS resolution times on the Mac, and about 17% degration on
Linux, when we don't probe.  It is likely that a few users
are greatly impacted by this, and are able to move the average
of the whole user population.


BUG=12754
BUG=25680
r=wtc
Review URL: http://codereview.chromium.org/1210003

git-svn-id: http://src.chromium.org/svn/trunk/src@42544 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 7faffcad3c449b20bde64562ae1d2dfdb62fd51b