Mac: Initial Geolocation UI.
For now, an in-progress version of the bubble/omnibox icon. The bubble shows up, but doesn't do much yet. The two buttons are disabled because they both don't do anything yet. Since this is behind a flag, I think checking in work-in-progress UI is fine.

Parts of this CL were merged in from http://codereview.chromium.org/1369002 . Ignore the files from that CL.

http://imgur.com/tWJhg

BUG=11246
TEST=Start chrome with --enable-geolocation, go to http://maxheapsize.com/static/html5geolocationdemo.html , click allow or deny, observe that an icon appears. Click icon, a bubble appears.

Review URL: http://codereview.chromium.org/1345005

git-svn-id: http://src.chromium.org/svn/trunk/src@42756 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 671b0ff8df7e665f62cdcfd58563f408343252b8