Split GpuProcessHost into GpuProcessHostUIShim, which runs on the UI thread, and GpuProcessHost, which now runs on the IO thread and derives from ChildProcessHost. This split was necessary in order to service synchronous messages from the renderer process. Moved message handlers for GPU messages from renderer to browser from BrowserRenderProcessHost to ResourceMessageFilter.
Stopped sending multiple ViewHostMsg_EstablishGpuChannel messages from
the same renderer if the connection was already established. Resetting
the channel was causing failures in Send, and every other page reload
containing WebGL content to fail. This cleanup will allow further
simplification in the GPU process, but this is being left for a
subsequent CL.

Fixed bug in sandboxing of GPU process. Fixed latent bugs in cleanup
code in GpuChannel and GpuChannelHost. Fixed crashes in
ChildProcessHost if resource_dispatcher_host_ was NULL. Fixed apparent
latent race conditions in creation of BackingStoreProxy and
VideoLayerProxy.

With these changes, WebGL content is running in the sandbox on both
Mac and Windows. Linux support will be added in a following CL.

BUG=29120
TEST=ran WebGL demos on Mac and Windows

Review URL: http://codereview.chromium.org/1546001

git-svn-id: http://src.chromium.org/svn/trunk/src@43029 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e015788e239c9000bbaddeb4415a1fc7b06c7784