[GTTF] Unify polling interval in some of the polling functions in the UI test framework:
- WaitUntilCookieValue
- WaitUntilCookieNonEmpty
- WaitUntilJavaScriptCondition

This helps reduce number of arbitrary "magic timeouts" which are different
in each test.

We don't remove the test timeout parameter, because it varies depending on test case,
and for now we don't have a better solution. Trying a "naive" way leads to hangs
on flaky tests.

TEST=UI-based tests
BUG=none

Review URL: http://codereview.chromium.org/1557001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@43062 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 60d241161cf132a58b772a499212daf4b2661892