Revert 43253 - 1. Create a new sandbox type which allows access to Unix sockets in the Mac renderer sandbox to support running Native Client. 2. Put the Native Client sel_ldr (which contains the user's untrusted code) into a new Mac sandbox type. 3. Open /dev/random in SandboxWarmup(). 4. Remove the "nosandbox" flag when running Mac tests.
See http://codereview.chromium.org/1234003/show which was reverted because of
problems on Mac 10.6. This change is identical except for the ";NACL" lines
in the *.sb files. I've removed the 10.6specific sandbox commands and used
the generic commands that work on 10.5 and 10.6. I will work on adding the
10.6specific commands in a different change list.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=327
TEST=nacl_ui_tests still pass while running in the sandbox.

Review URL: http://codereview.chromium.org/1525005

TBR=msneck@google.com
Review URL: http://codereview.chromium.org/1572009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@43259 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 204ba92b2dd48900d2f5ad6da8e0e284c30b2c0e