In order for CF to support FF3.6, the code that checks for privilege mode needs to be updated since it currently depends on APIs that have been deprecated.  The new method for determining if CF is running in privilege mode is to verify that the URL of the page that instantiates CF is a "chrome:" URL.
An advantage to this change is the code should now work in all version of
Firefox.  Previously, the code was different for each of versions 3.0, 3.5, and
now 3.6.

Note that the API change to npapi::GetStringIdentifiers() was necessary to get
around a problem where xpcom defines the same typedefs as chrome base, but with
different types.  If there is a better solution, please let me know.

TEST=see existing unit tests
BUG=none

Review URL: http://codereview.chromium.org/1535002

git-svn-id: http://src.chromium.org/svn/trunk/src@43311 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0f9bef783e7feffc23011db9ed011adce99c350c