Notify all active login prompts when one login prompt is submitted. This allows identical login prompts to dismiss themselves.
This was originally submitted as r42770 but had to be rolled back because I
copied from a FLAKY test. Now that r43266 fixed the flakiness, this should be
good to go.

The code is exactly the same as r42770 with one exception. I added a
WaitForNavigation guard to the tab the other tab in the new test cases. This
could have also caused flakiness in the initial revision.

BUG=8914
TEST=ui_tests --gtest_filter=LoginPromptTest.*

Review URL: http://codereview.chromium.org/1528012

git-svn-id: http://src.chromium.org/svn/trunk/src@43374 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 98d7c942afd07230e1fce0323ca021337b55ffbc