AutoFillDialogController leak fixed.
The AutoFillDialogController was leaking view controllers when dynamically adding
and removing AutoFillAddressViewController and AutoFillCreditCardViewController 
views.  Ownership of the controllers was not getting relinquished to the
NSArray containers correctly.

New checks in the unit tests now explicitly check that the controllers have a single owner.

BUG=39720
TEST=AutoFillDialogControllerTest.DeleteAddress, AutoFillDialogControllerTest.DeleteCreditCard
TBR=avi@chromium.org

Review URL: http://codereview.chromium.org/1551011

git-svn-id: http://src.chromium.org/svn/trunk/src@43432 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 172e73f5f548691ff84ba87f9a737a5d8bd68f0c