Fix regression for browser action container width on Toolkit Views.
Once we introduced the model for keeping track of the number of
visible icons, there was a race that was introduced where the 
toolbar could get constructed before the model knew about any
icons, so on startup the width of the container was 0.

This CL hooks up the newly added ModelLoaded notification and
sets the width on receiving that notification, if the width
hasn't been set already.

BUG=40329
TEST=None

Review URL: http://codereview.chromium.org/1547021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@43635 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f0c3b9901c569e34941487df9542e6e19e4ed737