Revert 43826 - HostResolver now adds AI_CANONNAME to the hint flags if a requester needs the information.
Requests which want the canonical name should be treated differently from requests that do not for the same host in both the HostCache as well as in the HostResolver when combining multiple outstanding requests into a job.

The motivation for this is that Kerberos SPN's for a web server are typically generated using the canonical name of the server rather than a DNS alias (both Firefox and IE have this behavior).

BUG=29862
TEST=net_unittests

Review URL: http://codereview.chromium.org/1566012

TBR=cbentzel@chromium.org
Review URL: http://codereview.chromium.org/1629005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@43833 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 79793a8f1f1e13aae6afe873d59c5ba4f16badbd