Don't bail early on extensions update.
When parsing the extension update manifest results, we bail out of there was an
error from any of the <app> tags in the xml. With the changes from

http://src.chromium.org/viewvc/chrome?view=rev&revision=42899

we now can get a "error-unknownApplication" for ids that Omaha doesn't know
about, so we need to not have that stop other updates.


BUG=40519
TEST=Install an extension you packaged yourself that doesn't have an 
update_url specified, and install an older verison of one in the gallery. When
you do an "update now" you should get the update for the gallery one.


Review URL: http://codereview.chromium.org/1633002

git-svn-id: http://src.chromium.org/svn/trunk/src@43891 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a3383d6a64c4ee937a78e628219795a2969fb8bc