Move GetExtension() from ExtensionFunctionDispatcher to ExtensionFunction.
EFD's can expire during the processing of AsyncExtensionFunctions, so this change enables them to retain access to their extension to complete their work.

This fixes a bug in executeScript where when a file source was used, but the call was made from a popup which which had closed, by the time the file source was loaded, the EFD had been destroyed. Go boom.

BUG=32431
TEST=All tests should pass

Review URL: http://codereview.chromium.org/1549026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@43999 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 851479d3717d893f6d9c28ba668096dd91a84c23