Send real clipping information to Mac plugins
Inform plugins of the actual clipping, so that they know not to try to invalidate off-screen areas (which in the case of Silverlight can cause it to stop drawing). We still always say that the window is the size of the plugin, and that the plugin is at (0,0), which is the part that actually matters for transforms and drawing location.

BUG=41323
TEST=Plugins scrolled partially off-screen should still draw correctly. Netflix video should keep playing even when the screen is smaller than the player.

Review URL: http://codereview.chromium.org/1558042

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@44396 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4afe2702b83e0f2b4f0fecdc4ac613f69a58a482