Send real clipping information to Mac plugins
Inform plugins of the actual clipping, so that they know not to try to invalidate off-screen areas (which in the case of Silverlight can cause it to stop drawing). We still always say that the window is the size of the plugin, and that the plugin is at (0,0), which is the part that actually matters for transforms and drawing location.

BUG=41323
TEST=Plugins scrolled partially off-screen should still draw correctly. Netflix video should keep playing even when the screen is smaller than the player.

Review URL: http://codereview.chromium.org/1558042

git-svn-id: http://src.chromium.org/svn/trunk/src@44396 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e96cbbfc4206fdd82296ccdfcdfbf6562803a7a3