Too many audio hangs Leopard
BUG=30242
TEST=Go to this page:
http://weston.ruter.net/projects/test-cases/chrome-mp3-audio-crash/
Click play and it won't hang.

Opening (not) too many audio streams in Leopard will cause a
deadlock in AudioQueue, freezing the browser process. Unfortunately
there isn't a way to detect that in the AudioQueue API. So this
change limits the number of streams allowed.

The limit for Leopard is set to 15 and other OSes are 50.

Review URL: http://codereview.chromium.org/1559024

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@44419 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3769f9047af3b59e5a031dbae1ae7bf1a744604a