Start up extension event routers later in BrowserMain, so bookmarks is loaded after bookmark import in first run.
I am not familiar with the extension code. It seems currently
ExtensionHistoryEventRouter and ExtensionAccessibilityEventRouter
are always initialized while ExtensionBrowserEventRouter
and ExtensionBookmarkEventRouter are initialized on demand. Since
we have a build in bookmark manager extension, I start all event
routers by default in this change. Let me know if this is an
issue.

BUG=40144
TEST=chrome.exe --first-run

Review URL: http://codereview.chromium.org/1578032

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@44668 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 6c2587df63477df7e29fac6746c04240b2acbb66