Rejigger how we format the display of bytes.
(Note that this does *not* touch the mega/mebi issue, despite my
recent discussion of it.)

What had been bugging me was that when downloading a file, you'd
see it count up 10.7, 10.8, 10.9, *11*, 11.1.  That is, previously
when we rounded to displaying a trailing zero, we'd clip it off,
which would make a download display like "11.1 / 50.2MB" jiggle to
the left every time the trailing digit lined up.

Now we try to always display three digits, which means we count up
like 98.8, 98.9, *99.0*, 99.1, ... 99.9, 100, ... 101.

This code is after fiddling the test into the shape I wanted it
to be, then making the code produce that behavior.

TEST=covered by unit test, updated with new expected behavior

Review URL: http://codereview.chromium.org/1558025

git-svn-id: http://src.chromium.org/svn/trunk/src@44671 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0655ba9bfdfdda131f3b76a38acf91abcfd2c00f