[Mac] Centralize hack to make tests work with AutocompleteClassifier.
A recent change added AutocompleteClassifier, which had to be hacked
into the testing profile to make tests with real location bars work.
Unfortunately, more code is tripping over this, so I'm moving the hack
to a central location rather than adding it case-by-case.

Issue 39725 is about removing the need for this.

BUG=39725
TEST=trybots.

Review URL: http://codereview.chromium.org/1618028

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@44775 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f8543d782a31e4f8b3527b83777eba34a9335aa7