Make Mac plugin _setWindowNumber: interception more focused
We interpose _setWindowNumber: because at the point where we intercept modal dialogs the window ID (which we need to pass to the browser process) hasn't been set yet. However, it also causes some false-positives. This narrows our _setWindowNumber: handling so that it only happens for the next call after we've caught an explicit window-showing call too early to get a window number.

BUG=41936
TEST=Plugin-opened windows (Flash full-screen, Gmail attachments, etc.) should continue to change focus correctly.

Review URL: http://codereview.chromium.org/1727001

git-svn-id: http://src.chromium.org/svn/trunk/src@45105 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 0b2fbfe4e5754e2d5480d211b49b3b03f2b17c86