Reland r45028: Factor out reading and writing of preferences into |PrefStore|.
In order to implement platform-specific policies, reading and writing preferences needs to be abstracted from the |PrefService|. The interface for that is now |PrefStore|, with an implementation |JsonPrefStore|, which stores the pref data in a JSON file. There is another implementation, |DummyPrefStore|, which doesn't store any persistent preferences, and is currently used for tests.

Most of the changes are for using the new interface, which is |new PrefService(new JsonPrefStore(filename))| instead of |new PrefService(filename)|.

BUG=40259
TEST=PrefServiceTest.*:PrefServiceSetValueTest.*:PrefMemberTest.*:JsonPrefStoreTest.*

Review URL: http://codereview.chromium.org/1687001

git-svn-id: http://src.chromium.org/svn/trunk/src@45168 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 286a310f65593c54aed8857e9ee12fa2966dc1a2