Reland r45028: Factor out reading and writing of preferences into |PrefStore|.
In order to implement platform-specific policies, reading and writing preferences needs to be abstracted from the |PrefService|. The interface for that is now |PrefStore|, with an implementation |JsonPrefStore|, which stores the pref data in a JSON file. There is another implementation, |DummyPrefStore|, which doesn't store any persistent preferences, and is currently used for tests.

Most of the changes are for using the new interface, which is |new PrefService(new JsonPrefStore(filename))| instead of |new PrefService(filename)|.

BUG=40259
TEST=PrefServiceTest.*:PrefServiceSetValueTest.*:PrefMemberTest.*:JsonPrefStoreTest.*

Review URL: http://codereview.chromium.org/1687001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@45168 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7e0db23962472f6fe795d76d0773ca1c9ab9ac77