Increase sleep timeout for valgrind UI tests.
DownloadTest.CloseNewTab2 is failing (ASSERT failure, not memory leak) under valgrind because of a timeout.

This patch increases the sleep timeout for UI tests to the same value as the action timeout (when running the test normally, the default value for the sleep timeout is also the action timeout).

BUG=none
TEST=DownloadTest.CloseNewTab2 (run under valgrind)

Review URL: http://codereview.chromium.org/1725003

git-svn-id: http://src.chromium.org/svn/trunk/src@45183 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a9bce2a0fc3d1080fc1278d90d0dcd6a3929d2d1