Windows bots build and use OSMesa
This is a software GL renderer that can be used by bots that do not have access to native GL. It will be used for running tests that use GL, such as Pepper 3D, WebGL and accelerated compositing. Only working on Windows so far.

TEST=trybots
BUG=none

Review URL: http://codereview.chromium.org/1740001

git-svn-id: http://src.chromium.org/svn/trunk/src@45241 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5ad9cfdb12e5b6d0eb9443937b490e96d6bc14e6