Refactor the split pane element into a splitter element.
The reason for the refactoring is so that the splitter can be used in cases where there are multiple columns.

This also adds a workaround for https://bugs.webkit.org/show_bug.cgi?id=32230 which causes the splitter to not work correctly when the page is zoomed.

BUG=42282, 42290
TEST=Open the bookmark manager. Zoom the page. Then drag the splitter.

Review URL: http://codereview.chromium.org/1763004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@45356 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 99e6db8bd9956e50837db2f7c9501f1fbe0a000e