Add an --allow-external-pages switch to test_shell.
Normally test_shell will not let you load an external URL (like
"www.google.com") when running in --layout-tests mode. This is a good thing
because it keeps us from writing tests that might have unpredictable results
if the external URL changes.

However, it would be cool to be able to use the --layout-test infrastructure
to capture the rendered/painted page (outside of the actual layout tests run)
for QA purposes. This change adds an --allow-external-pages switch to
override the check described above.

This is kind of a short-term fix. A better fix would be to add a "--headless"
flag that loaded and rendered the page and dumped the whole thing, entirely
separate from --layout-test mode. I will add that soon.

    BUG=none
    TEST=none
    R=dglazkov

Review URL: http://codereview.chromium.org/1748005

git-svn-id: http://src.chromium.org/svn/trunk/src@45403 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 6bc9c941257de6eac2210b52b679e04adbd2bece