Bookmark manager: Fix issue with mouse events on the scrollbar as well as clicking on empty areas with keyboard modifiers.
In case where a mouse event originates on the scrollbar we exit early.

On Window, if the user clicks on an empty area and holds down a key modifier we should do nothing.
On Mac, we always clear the selection when clicking on an empty area, no matter if any keys are pressed.

BUG=42371
TEST=Select some items in the list. Scrolling the list using the mouse and scrollbar should not change the selection.
     Shift and/or ctrl clicking on an empty area should clear selection on mac and do nothing on windows.


Review URL: http://codereview.chromium.org/1694013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@45487 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0d4df9bdca141f455d54f966e6f2259216a4e7c9