Bookmark manager: Fix issue with mouse events on the scrollbar as well as clicking on empty areas with keyboard modifiers.
In case where a mouse event originates on the scrollbar we exit early.

On Window, if the user clicks on an empty area and holds down a key modifier we should do nothing.
On Mac, we always clear the selection when clicking on an empty area, no matter if any keys are pressed.

BUG=42371
TEST=Select some items in the list. Scrolling the list using the mouse and scrollbar should not change the selection.
     Shift and/or ctrl clicking on an empty area should clear selection on mac and do nothing on windows.


Review URL: http://codereview.chromium.org/1694013

git-svn-id: http://src.chromium.org/svn/trunk/src@45487 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 11f11e63031e1dff98655e2ec762c464c1528ad4