Mac: Fix use-after-free in translate infobar options menu.
The "About Translate" command opens a new tab which closes the infobar, we where trying to update the options menu which acccessed an already released object.

The fix is to match Windows behavior and only rebuild the options menu when clicked.

Thanks to Rohit for diagnosing the problem!

BUG=42030
TEST=repro in bug description.

Review URL: http://codereview.chromium.org/1756006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@45553 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: beb3879a1de29882d4b2bdc055ede79862d62a45