Two changes to extension app icons:
1) When one of the mini tabs is an app (nano) tab, we
now increase the margin between the mini tabs and the
regular tabs by a few pixels.

This is because with the large icons for nano tabs, it
looks a little squished against the regular tabs in the
tabstrip.

2) When dragging a nano tab within the tab strip, the
background "handle" would grow to be as large as a
regular tab (while dragging) due to the fact that
UpdateData (that DraggedTabView calls) was not
transferring the app tab bit to the new Tab.

BUG=None
TEST=Visual inspection.

Review URL: http://codereview.chromium.org/1700016

git-svn-id: http://src.chromium.org/svn/trunk/src@45664 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 8bab41b81ca8dd8693bf28bdbd2c776fd5704e4d