Adding ExtensionPrefs methods for storing update-when-idle data.
This includes adding a few methods to ExtensionPrefs that I'm going to need to
use inside ExtensionUpdater for implementing the "do updates at idle" feature.
Instead of adding more call-through stubs to the interface that 
ExtensionsService providers to the ExtensionUpdater, I instead decided it was 
time to do some refactoring to expose a ExtensionPrefs getter in the interface.

TEST=(Should be covered by unit/browser tests)
BUG=37971


Review URL: http://codereview.chromium.org/1695018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@45738 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 73380bb3faceae5801441203467754b428ddcd06