linux: reset GDK_NATIVE_WINDOWS, fixing clicking on flash again
Even though we set this environment variable, GTK clears it after initializing
itself from it, with the intent of isolating the variable's value to just
this process.

If nspluginwrapper is involved, though, we want to propagate the variable to
its child as well.  So we must set the variable both before initialization
(to initialize it for the current process) as well as after (to initialize
it for the child process, if any).

Review URL: http://codereview.chromium.org/1789008

git-svn-id: http://src.chromium.org/svn/trunk/src@45886 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 449b1631c93bfc7bd287e211b900e7f8e7723687