First cut at window affinity for apps. Window affinity means that app URLs tend to open in the window associated with them, and non-app URLs tend to not open in app windows.
After playing with this first attempt, I find it is unsatisfying because spreadsheets.google.com doesn't open in the docs app (different origin) and logging out of gmail directs you back to the browser.

So I'm going to fix that by adding a browse extent, separately.

But I still think this is a worthwhile stepping stone to check in.

Review URL: http://codereview.chromium.org/1732015

git-svn-id: http://src.chromium.org/svn/trunk/src@45894 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d3321b0d9130a2f906b9eab2a933ef528980d84d