Fix Dropping in the bookmark manager
Originally we were clearing the dragData member in a few places, including
experimental.bookmarkManager.onDrop.  This was triggered before the javascripts
drop event so when it got to that point to drop a valid bookmark, it had already
cleared itself out.  Now it just posts a task with 0ms delay to let the real
drop event handle the data before clearing it.  Also clear the data on mouse up,
which wasn't being handled in the case where you drag an element just a tiny
bit, and no action is performed.

BUG=none
TEST=Ensure you can still drop stuff in the bmm

Review URL: http://codereview.chromium.org/1780015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@46006 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a7c57ec35c10f5dcacfa24cbb21caf67fa3cb321