The windowed flash plugin has a bug which occurs when the plugin enters fullscreen mode. It basically captures the mouse on WM_LBUTTONDOWN and does not release capture correctly causing it to stop receiving subsequent mouse events.
This problem is also seen in Safari where there is code to handle this in the wndproc. However the plugin subclasses the window in
WM_LBUTTONDOWN before entering full screen. As a result Safari does not receive the WM_LBUTTONUP message. 

To workaround this issue we use a per thread mouse hook to capture and release the mouse. This hook is only installed
on a quirk. 

This bug does not occur in Firefox and opera. Firefox has code similar to Safari where it captures the mouse on
WM_LBUTTONDOWN and releases it on WM_LBUTTONUP. It could well be a bug in the flash plugin, which only occurs in Safari and Chrome.

Fixes http://code.google.com/p/chromium/issues/detail?id=28025

Bug=28025


Review URL: http://codereview.chromium.org/1718023

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@46031 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 11b19d4ae64061e22d854a907408c02aa9120926