Allow authenticated installs from the disk image.
When launching from a read-only disk image, if no installed copy exists in
/Applications, and the user chooses to install from the disk image:

If it appears that there isn't permission to write to /Applications, prompt
the user to authenticate as an administrator.  If authentication is
successful, copy the application as root, set its permissions, and promote
Keystone to use a system ticket.

BUG=40265
TEST=Remove the application from /Applications and test, as both an admin
     user and a normal user, installation directly from the disk image.  When
     installation occurs after authentication, Chrome should be on a system
     ticket.
     Also, make sure that normal in-app Keystone promotion continues to work
     as expected.
Review URL: http://codereview.chromium.org/1792013

git-svn-id: http://src.chromium.org/svn/trunk/src@46091 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: d6549d9ef0a6e8a0a6749a8774e5751e2ae1e209