Fix a potential crash in ChromeOS options dialog.
This is because the gtk-based options page is based on NativeViewHost
which does not provide CanClose function. The fix bypasses the check of
CanClose for gtk-base options pages since this seems not supported.
The crash does not happen now because OptionsWindowView does not set
accept_ flag and thus CanClose is not called. However, we might want to
patch this just in case.

BUG=none
TEST=Verify Chrome does not crash on ChromeOs when user presses "ESC" on gtk-based options page.

Review URL: http://codereview.chromium.org/1825001

git-svn-id: http://src.chromium.org/svn/trunk/src@46127 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 5102feb541fc309b34c963790c9a00b80c6c85e1