Initial support for web-extent background pages.
This patch adds a new RVH container: BackgroundContents. The idea is that apps can open a live web-page as a "background" page using window.open('<url>', '<name>', 'background');

If 'background' is specified and the opener is within the app's extent, a BackgroundContents will be used. Otherwise, the 'background' feature is ignored and it is treated as a regular popup call.

Note that as of this patch the following are explicitly not-yet addressed:

1) Session storage for BackgroundContents
2) SSL (or other failures) requiring UI
3) Javascript messages (alert, etc...)
4) Session restore

TEST=All tests should pass
BUG=41275

Review URL: http://codereview.chromium.org/1734014

git-svn-id: http://src.chromium.org/svn/trunk/src@46544 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 2aad71c3020bbaa9e601829ecb94f039a24dbf81