This CL adds the option to support buffers that are bound to more than one target.
We had originally decided that this would be an disallowed.

glBindBuffer(GL_ARRAY_BUFFER, some_buffer_id);
glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, some_buffer_id);

Unfortunately the GLES2 conformance tests require this
behavior.

This CL makes this behavior possible but it has to be
turned on because it means all buffers have to have CPU
side backing (not just ELEMENT_ARRAY_BUFFERs) and it
means a slowdown for dynamic vertex related stuff
like skinning because an extra copy has to be made
every time they are updated.

So, by default it will be off but we'll come up with
some way to turn it on. Al's suggestion is to use
eglCreateContext with a custom intialization attribute.

My #1 question: What about the name of the command,
the name of the constant and the filename the constant
is in. Any thoughts?

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/1992008

git-svn-id: http://src.chromium.org/svn/trunk/src@46756 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: 625d5041085af94c0cf48bc300116b251365f4ea