Fix possible deadlock in PluginChannel.
This occurs when the renderer sends an async message with the unblock flag, and then a sync message right after.  If the plugin process just made a sync (with no unblock) call to the renderer, it'll dispatch the first message, and if that leads to a sync call to the renderer, then the unblock flag won't get sent and a deadlock occurs.

BUG=43617
Review URL: http://codereview.chromium.org/2045012

git-svn-id: http://src.chromium.org/svn/trunk/src@47063 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: e387d95a789eba55311ca029d4746ebc8d401082