Add --host-rules support.
The format for --host-rules is identical to --host-resolver-rules.
The difference is that --host-rules affects the endpoint of the HttpNetworkTransaction, not just the host resolver.  So, this means the host passed to the host resolver and the TCP connect(), the tunnel CONNECT, and the SOCKS connect will be different.

Review URL: http://codereview.chromium.org/2057007

git-svn-id: http://src.chromium.org/svn/trunk/src@47083 4ff67af0-8c30-449e-8e8b-ad334ec8d88c

Former-commit-id: a4feb8c45b100a734ddf2ebe60ec0c233ce8998a